restructure sectionView, unify modal state with url
