--- frontend/src/app/Catalog/CatalogFilters/index.ts
@@ -1,3 +1,0 @@
-import CatalogFilters from './CatalogFilters';
-
-export default CatalogFilters;
