--- frontend/src/app/Catalog/CatalogList/index.ts
@@ -1,3 +1,0 @@
-import CatalogList from './CatalogList';
-
-export default CatalogList;
